Fjernet OpplaringKategoriseringResponse.Alternativ.Gruppe (#234)
* Fjernet OpplaringKategoriseringResponse.Alternativ.Gruppe
* Lagt til representerer som enum
* SertifiseringValg.id: Int -> Long

+++ amt-deltaker/src/main/kotlin/no/nav/amt/deltaker/repository/SertifiseringValgRepository.kt
@@ -55,7 +55,7 @@ object SertifiseringValgRepository {
                     queryOf(sql, deltakerlisteId)
                         .map { row ->
                             SertifiseringValg(
-                                id = row.int("sertifisering_id"),
+                                id = row.long("sertifisering_id"),
                                 navn = row.string("sertifisering_navn"),
                             )
                         }.asList,
